Move make_whole tests to test_mdamath.py
Relocated the TestMakeWhole test class from test_util.py to test_mdamath.py for better organization and relevance. This change helps keep tests for mdamath functions within their dedicated test module.
