+++ testsuite/MDAnalysisTests/lib/test_util.py
@@ -43,7 +43,6 @@ from itertools import combinations_with_replacement as comb_wr
 
 import MDAnalysis as mda
 import MDAnalysis.lib.util as util
-import MDAnalysis.lib.mdamath as mdamath
 from MDAnalysis.lib.util import (
     cached,
     static_variables,
@@ -228,272 +227,6 @@ class TestFilename(object):
             pytest.fail("fn and ns are different")
         assert str(fn) == self.filename2
         assert ns.name == self.filename2
-
-
-class TestMakeWhole(object):
-    """Set up a simple system:
-
-    +-----------+
-    |           |
-    | 6       3 | 6
-    | !       ! | !
-    |-5-8   1-2-|-5-8
-    | !       ! | !
-    | 7       4 | 7
-    |           |
-    +-----------+
-    """
-
-    prec = 5
-
-    @pytest.fixture()
-    def universe(self):
-        universe = mda.Universe(Make_Whole)
-        bondlist = [(0, 1), (1, 2), (1, 3), (1, 4), (4, 5), (4, 6), (4, 7)]
-        universe.add_TopologyAttr(Bonds(bondlist))
-        return universe
-
-    def test_return_value(self, universe):
-        ag = universe.residues[0].atoms
-        orig_pos = ag.positions.copy()
-        retval = mdamath.make_whole(ag)
-        assert retval.dtype == np.float32
-        assert_array_equal(ag.positions, retval)
-        assert np.any(ag.positions != orig_pos)
-
-    def test_single_atom_no_bonds(self):
-        # Call make_whole on single atom with no bonds, shouldn't move
-        u = mda.Universe(Make_Whole)
-        # Atom0 is isolated
-        bondlist = [(1, 2), (1, 3), (1, 4), (4, 5), (4, 6), (4, 7)]
-        u.add_TopologyAttr(Bonds(bondlist))
-
-        ag = u.atoms[[0]]
-        refpos = ag.positions.copy()
-        mdamath.make_whole(ag)
-
-        assert_array_equal(ag.positions, refpos)  # must be untouched
-
-    def test_empty_ag(self, universe):
-        ag = mda.AtomGroup([], universe)
-        retval = mdamath.make_whole(ag)
-        assert retval.dtype == np.float32
-        assert_array_equal(retval, np.empty((0, 3), dtype=np.float32))
-
-    def test_scrambled_ag(self, universe):
-        # if order of atomgroup is mixed
-        ag = universe.atoms[[1, 3, 2, 4, 0, 6, 5, 7]]
-
-        mdamath.make_whole(ag)
-
-        # artificial system which uses 1nm bonds, so
-        # largest bond should be 20A
-        assert ag.bonds.values().max() < 20.1
-
-    def test_out_of_place(self, universe):
-        ag = universe.residues[0].atoms
-        orig_pos = ag.positions.copy()
-        mdamath.make_whole(ag, inplace=False)
-        # positions must be untouched:
-        assert_array_equal(ag.positions, orig_pos)
-
-    def test_double_precision_box(self):
-        # This test could in principle be removed since PR #2213
-        # universe with double precision box containing a 2-atom molecule
-        # broken accross a corner:
-        u = mda.Universe.empty(
-            n_atoms=2,
-            n_residues=1,
-            n_segments=1,
-            atom_resindex=[0, 0],
-            residue_segindex=[0],
-            trajectory=True,
-            velocities=False,
-            forces=False,
-        )
-        ts = u.trajectory.ts
-        ts.frame = 0
-        ts.dimensions = [10, 10, 10, 90, 90, 90]
-        # assert ts.dimensions.dtype == np.float64
-        # not applicable since #2213
-        ts.positions = np.array(
-            [
-                [1, 1, 1],
-                [9, 9, 9],
-            ],
-            dtype=np.float32,
-        )
-        u.add_TopologyAttr(Bonds([(0, 1)]))
-        mdamath.make_whole(u.atoms)
-        assert_array_almost_equal(
-            u.atoms.positions,
-            np.array(
-                [
-                    [1, 1, 1],
-                    [-1, -1, -1],
-                ],
-                dtype=np.float32,
-            ),
-        )
-
-    @staticmethod
-    @pytest.fixture()
-    def ag(universe):
-        return universe.residues[0].atoms
-
-    def test_no_bonds(self):
-        # NoData caused by no bonds
-        universe = mda.Universe(Make_Whole)
-        ag = universe.residues[0].atoms
-        with pytest.raises(NoDataError):
-            mdamath.make_whole(ag)
-
-    def test_zero_box_size(self, universe, ag):
-        universe.dimensions = [0.0, 0.0, 0.0, 90.0, 90.0, 90.0]
-        with pytest.raises(ValueError):
-            mdamath.make_whole(ag)
-
-    def test_wrong_reference_atom(self, universe, ag):
-        # Reference atom not in atomgroup
-        with pytest.raises(ValueError):
-            mdamath.make_whole(ag, reference_atom=universe.atoms[-1])
-
-    def test_impossible_solve(self, universe):
-        # check that the algorithm sees the bad walk
-        with pytest.raises(ValueError):
-            mdamath.make_whole(universe.atoms)
-
-    def test_solve_1(self, universe, ag):
-        # regular usage of function
-
-        refpos = universe.atoms[:4].positions.copy()
-
-        mdamath.make_whole(ag)
-
-        assert_array_almost_equal(universe.atoms[:4].positions, refpos)
-        assert_array_almost_equal(
-            universe.atoms[4].position,
-            np.array([110.0, 50.0, 0.0]),
-            decimal=self.prec,
-        )
-        assert_array_almost_equal(
-            universe.atoms[5].position,
-            np.array([110.0, 60.0, 0.0]),
-            decimal=self.prec,
-        )
-        assert_array_almost_equal(
-            universe.atoms[6].position,
-            np.array([110.0, 40.0, 0.0]),
-            decimal=self.prec,
-        )
-        assert_array_almost_equal(
-            universe.atoms[7].position,
-            np.array([120.0, 50.0, 0.0]),
-            decimal=self.prec,
-        )
-
-    def test_solve_2(self, universe, ag):
-        # use but specify the center atom
-
-        refpos = universe.atoms[4:8].positions.copy()
-
-        mdamath.make_whole(ag, reference_atom=universe.residues[0].atoms[4])
-
-        assert_array_almost_equal(universe.atoms[4:8].positions, refpos)
-        assert_array_almost_equal(
-            universe.atoms[0].position,
-            np.array([-20.0, 50.0, 0.0]),
-            decimal=self.prec,
-        )
-        assert_array_almost_equal(
-            universe.atoms[1].position,
-            np.array([-10.0, 50.0, 0.0]),
-            decimal=self.prec,
-        )
-        assert_array_almost_equal(
-            universe.atoms[2].position,
-            np.array([-10.0, 60.0, 0.0]),
-            decimal=self.prec,
-        )
-        assert_array_almost_equal(
-            universe.atoms[3].position,
-            np.array([-10.0, 40.0, 0.0]),
-            decimal=self.prec,
-        )
-
-    def test_solve_3(self, universe):
-        # put in a chunk that doesn't need any work
-
-        refpos = universe.atoms[:1].positions.copy()
-
-        mdamath.make_whole(universe.atoms[:1])
-
-        assert_array_almost_equal(universe.atoms[:1].positions, refpos)
-
-    def test_solve_4(self, universe):
-        # Put in only some of a fragment,
-        # check that not everything gets moved
-
-        chunk = universe.atoms[:7]
-        refpos = universe.atoms[7].position.copy()
-
-        mdamath.make_whole(chunk)
-
-        assert_array_almost_equal(universe.atoms[7].position, refpos)
-        assert_array_almost_equal(
-            universe.atoms[4].position, np.array([110.0, 50.0, 0.0])
-        )
-        assert_array_almost_equal(
-            universe.atoms[5].position, np.array([110.0, 60.0, 0.0])
-        )
-        assert_array_almost_equal(
-            universe.atoms[6].position, np.array([110.0, 40.0, 0.0])
-        )
-
-    def test_double_frag_short_bonds(self, universe, ag):
-        # previous bug where if two fragments are given
-        # but all bonds were short, the algorithm didn't
-        # complain
-        mdamath.make_whole(ag)
-        with pytest.raises(ValueError):
-            mdamath.make_whole(universe.atoms)
-
-    def test_make_whole_triclinic(self):
-        u = mda.Universe(TPR, GRO)
-        thing = u.select_atoms("not resname SOL NA+")
-        mdamath.make_whole(thing)
-
-        blengths = thing.bonds.values()
-
-        assert blengths.max() < 2.0
-
-    def test_make_whole_fullerene(self):
-        # lots of circular bonds as a nice pathological case
-        u = mda.Universe(fullerene)
-
-        bbox = u.atoms.bbox()
-        u.dimensions = np.r_[bbox[1] - bbox[0], [90] * 3]
-
-        blengths = u.atoms.bonds.values()
-        # kaboom
-        u.atoms[::2].translate([u.dimensions[0], -2 * u.dimensions[1], 0.0])
-        u.atoms[1::2].translate(
-            [0.0, 7 * u.dimensions[1], -5 * u.dimensions[2]]
-        )
-
-        mdamath.make_whole(u.atoms)
-
-        assert_array_almost_equal(
-            u.atoms.bonds.values(), blengths, decimal=self.prec
-        )
-
-    def test_make_whole_multiple_molecules(self):
-        u = mda.Universe(two_water_gro, guess_bonds=True)
-
-        for f in u.atoms.fragments:
-            mdamath.make_whole(f)
-
-        assert u.atoms.bonds.values().max() < 2.0
 
 
 class Class_with_Caches(object):
